+++ Sources/SuperwallKit/StoreKit/Coordinator/StoreKitCoordinator.swift
@@ -7,40 +7,8 @@
 
 import Foundation
 
-protocol TransactionChecker: AnyObject {
-  /// Gets and validates a transaction of a product.
-  func getAndValidateTransaction(
-    of productId: String,
-    since purchaseStartDate: Date
-  ) async throws -> StoreTransaction
-}
-
-protocol ProductPurchaser: AnyObject {
-  /// Purchases a product and returns its result.
-  func purchase(product: StoreProduct) async -> PurchaseResult
-}
-
-protocol ProductsFetcher: AnyObject {
-  /// Fetches a set of products from their identifiers.
-  func products(identifiers: Set<String>) async throws -> Set<StoreProduct>
-}
-
-protocol TransactionRestorer: AnyObject {
-  /// Restores purchases.
-  ///
-  /// - Returns: A boolean indicating whether the restore request succeeded or failed.
-  /// This doesn't mean that the user is now subscribed, just that there were no errors
-  /// obtaining the restored transactions
-  func restorePurchases() async -> Bool
-}
-
-protocol SubscriptionStatusChecker: AnyObject {
-  // TODO: Should this be async?
-  /// Determines the subscription status of the user.
-  func isSubscribed() -> Bool
-}
-
-// TODO: Add generics to make sure the correct StoreProduct is purchased rather than guards
+/// Coordinates: The purchasing, restoring and retrieving of products; the checking
+/// of transactions; and the determining of the user's subscription status.
 struct StoreKitCoordinator {
   /// Purchases the product.
   let productPurchaser: ProductPurchaser
@@ -70,32 +38,22 @@ struct StoreKitCoordinator {
     let hasDelegate = purchasingDelegateAdapter.hasDelegate
 
     if hasDelegate {
-      // If delegate exists, finishTransactions will always be false.
-      if #available(iOS 15.0, tvOS 15.0, watchOS 8.0, *) {
-        // TODO: Check this, we need to make sure that SK1 and SK2 don't interfere!
-        // TODO: Check that someone can purchase using RevenueCat in SK2 but restore purchases using our SK1 handler.
-        // Have to rely on the delegate to purchase because
-        self.productPurchaser = purchasingDelegateAdapter
-        self.productFetcher = ProductsFetcherSK1()
-        self.txnChecker = ProductPurchaserSK2(
-          storeKitManager: storeKitManager,
-          sessionEventsManager: sessionEventsManager,
-          factory: factory
-        )
-        self.txnRestorer = purchasingDelegateAdapter
-        self.subscriptionStatusHandler = purchasingDelegateAdapter
-
-      } else {
-        self.productPurchaser = purchasingDelegateAdapter
-        self.productFetcher = ProductsFetcherSK1()
-        self.txnChecker = ProductPurchaserSK1(
-          storeKitManager: storeKitManager,
-          sessionEventsManager: sessionEventsManager,
-          factory: factory
-        )
-        self.txnRestorer = purchasingDelegateAdapter
-        self.subscriptionStatusHandler = purchasingDelegateAdapter
-      }
+      // If delegate exists, finishTransactions will always be false and SK1 must be used.
+      // After testing, it turns out we can't use SK2 here because:
+      // 1. If there are unfinished StoreKit transaction updates on app open, only the first
+      //    listener is given the updates. This means that our listener can interfere with
+      //    the clients listener (if they're using SK2) and their transactions may not even show up!
+      // 2. If they use an SK1 listener, their listener could finish the txns before
+      //    ours. Which means that we will never see the unfinished transactions.
+      self.productPurchaser = purchasingDelegateAdapter
+      self.productFetcher = ProductsFetcherSK1()
+      self.txnChecker = ProductPurchaserSK1(
+        storeKitManager: storeKitManager,
+        sessionEventsManager: sessionEventsManager,
+        factory: factory
+      )
+      self.txnRestorer = purchasingDelegateAdapter
+      self.subscriptionStatusHandler = purchasingDelegateAdapter
     } else {
       if #available(iOS 15.0, tvOS 15.0, watchOS 8.0, *) {
         if finishTransactions {
@@ -122,7 +80,8 @@ struct StoreKitCoordinator {
           self.subscriptionStatusHandler = storeKitManager
         }
       } else {
-        // Regardless of finishing transactions.
+        // We have to use SK1 on <iOS 15, regardless of whether
+        // we finish transactions.
         let purchaser = ProductPurchaserSK1(
           storeKitManager: storeKitManager,
           sessionEventsManager: sessionEventsManager,
