+++ Sources/SuperwallKit/StoreKit/Products/Receipt Manager/Receipt Models/InAppPurchase.swift
