+++ Sources/SuperwallKit/StoreKit/Products/Receipt Manager/Receipt Models/InAppReceipt.swift
