+++ Sources/SuperwallKit/StoreKit/Products/Receipt Manager/Receipt Models/InAppReceiptAttribute.swift
