+++ Sources/SuperwallKit/StoreKit/Products/Receipt Manager/Receipt Models/InAppReceiptPayload.swift
