+++ Sources/SuperwallKit/StoreKit/Products/Receipt Manager/Receipt Models/InAppReceiptPayloadContainer.swift
